feat: remove some mle dependencies
