+++ core/src/authorization/decorators/current-player.decorator.ts
@@ -2,11 +2,7 @@ import type {ExecutionContext} from "@nestjs/common";
 import {createParamDecorator} from "@nestjs/common";
 import {GqlExecutionContext} from "@nestjs/graphql";
 
-import type {Player} from "$models";
-
 export const CurrentPlayer = createParamDecorator((data: unknown, context: ExecutionContext) => {
     const ctx = GqlExecutionContext.create(context);
-    const player = ctx.getContext().req.player as Player | undefined;
-    if (!player) throw new Error("Cannot locate Player from PlayerGuard");
-    return player;
+    return ctx.getContext().req.player;
 });
