+++ core/src/authorization/decorators/index.ts
@@ -1,0 +1,3 @@
+export * from "./actions.decorator";
+export * from "./current-member.decorator";
+export * from "./current-player.decorator";
