+++ core/src/database/franchise/player/player.repository.ts
@@ -1,4 +1,5 @@
 import {Injectable} from "@nestjs/common";
+import type {FindOneOptions} from "typeorm";
 import {DataSource} from "typeorm";
 
 import {ExtendedRepository} from "../../extended-repositories/repository";
@@ -8,5 +9,33 @@ import {Player} from "./player.model";
 export class PlayerRepository extends ExtendedRepository<Player> {
     constructor(readonly dataSource: DataSource) {
         super(Player, dataSource);
+    }
+
+    async getByOrganizationAndGame(
+        userId: number,
+        organizationId: number,
+        gameId: number,
+        options?: FindOneOptions<Player>,
+    ): Promise<Player> {
+        return this.get(
+            Object.assign(
+                {
+                    where: {
+                        member: {
+                            userId,
+                            organizationId,
+                        },
+                        skillGroup: {
+                            gameId,
+                        },
+                    },
+                    relations: {
+                        member: true,
+                        skillGroup: true,
+                    },
+                },
+                options,
+            ),
+        );
     }
 }
