+++ core/src/franchise/player/index.ts
@@ -1,3 +1,2 @@
-export * from "./current-player.decorator";
 export * from "./player.guard";
 export * from "./player.service";
