+++ core/src/mledb/index.ts
@@ -1,4 +1,3 @@
 export * from "./mledb-interface.module";
 export * from "./mledb-player";
-export * from "./mledb-player-account";
 export * from "./mledb-scrim";
