+++ core/src/mledb/mledb-interface.module.ts
@@ -2,17 +2,16 @@ import {forwardRef, Module} from "@nestjs/common";
 import {MatchmakingModule} from "@sprocketbot/common";
 
 import {DatabaseModule} from "../database";
+import {EloConnectorModule} from "../elo/elo-connector";
 import {FranchiseModule} from "../franchise";
 import {GameModule} from "../game";
-import {IdentityModule} from "../identity";
-import {OrganizationModule} from "../organization";
 import {SchedulingModule} from "../scheduling/scheduling.module";
 import {SprocketRatingModule} from "../sprocket-rating";
 import {UtilModule} from "../util/util.module";
 import {MledbMatchController} from "./mledb-match/mledb-match.controller";
 import {MledbMatchService} from "./mledb-match/mledb-match.service";
 import {MledbPlayerService} from "./mledb-player";
-import {MledbPlayerAccountService} from "./mledb-player-account";
+import {MledbPlayerResolver} from "./mledb-player/mledb-player.resolver";
 import {MledbFinalizationService} from "./mledb-scrim";
 
 @Module({
@@ -22,13 +21,12 @@ import {MledbFinalizationService} from "./mledb-scrim";
         MatchmakingModule,
         SprocketRatingModule,
         UtilModule,
+        EloConnectorModule,
         forwardRef(() => SchedulingModule),
         forwardRef(() => FranchiseModule),
-        forwardRef(() => IdentityModule),
-        forwardRef(() => OrganizationModule),
     ],
-    providers: [MledbPlayerService, MledbPlayerAccountService, MledbFinalizationService, MledbMatchService],
-    exports: [MledbMatchService, MledbPlayerService, MledbPlayerAccountService, MledbFinalizationService],
+    providers: [MledbPlayerService, MledbFinalizationService, MledbMatchService, MledbPlayerResolver],
+    exports: [MledbMatchService, MledbPlayerService, MledbFinalizationService],
     controllers: [MledbMatchController],
 })
 export class MledbInterfaceModule {}
