--- core/src/mledb/mledb-player-account/index.ts
@@ -1,1 +1,0 @@
-export * from "./mledb-player-account.service";
