--- core/src/mledb/mledb-player-account/mledb-player-account.service.ts
@@ -1,18 +1,0 @@
-import {Injectable} from "@nestjs/common";
-import {InjectRepository} from "@nestjs/typeorm";
-import type {FindManyOptions} from "typeorm";
-import {Repository} from "typeorm";
-
-import {MLE_PlayerAccount} from "$mledb";
-
-@Injectable()
-export class MledbPlayerAccountService {
-    constructor(
-        @InjectRepository(MLE_PlayerAccount)
-        private playerAccountRepository: Repository<MLE_PlayerAccount>,
-    ) {}
-
-    async getPlayerAccounts(query: FindManyOptions<MLE_PlayerAccount>): Promise<MLE_PlayerAccount[]> {
-        return this.playerAccountRepository.find(query);
-    }
-}
