+++ core/src/mledb/mledb-player/mledb-player.guard.ts
@@ -13,8 +13,13 @@ export class FormerPlayerScrimGuard implements CanActivate {
     async canActivate(context: ExecutionContext): Promise<boolean> {
         const ctx = GqlExecutionContext.create(context);
         const payload = ctx.getContext().req.user as JwtAuthPayload;
-        const mlePlayer = await this.mledbPlayerService.getMlePlayerBySprocketUser(payload.userId);
-        if (mlePlayer.teamName === "FP") throw new GraphQLError("User is a former player in MLE");
+
+        try {
+            const mlePlayer = await this.mledbPlayerService.getMlePlayerBySprocketUser(payload.userId);
+            if (mlePlayer.teamName === "FP") throw new GraphQLError("User is a former player in MLE");
+        } catch {
+            return true;
+        }
 
         return true;
     }
